mlx5: share Netlink commands
Move Netlink mechanism and its dependecies from net/mlx5 to common/mlx5
in order to be ready to use by other mlx5 drivers.

The dependencies are BITFIELD defines, the ppc64 compilation workaround
for bool type and the function mlx5_translate_port_name.

Update build mechanism accordingly.
