+++ drivers/common/mlx5/mlx5_nl.h
@@ -7,25 +7,10 @@
 
 #include <linux/netlink.h>
 
+#include <rte_ether.h>
 
-/* Recognized Infiniband device physical port name types. */
-enum mlx5_nl_phys_port_name_type {
-	MLX5_PHYS_PORT_NAME_TYPE_NOTSET = 0, /* Not set. */
-	MLX5_PHYS_PORT_NAME_TYPE_LEGACY, /* before kernel ver < 5.0 */
-	MLX5_PHYS_PORT_NAME_TYPE_UPLINK, /* p0, kernel ver >= 5.0 */
-	MLX5_PHYS_PORT_NAME_TYPE_PFVF, /* pf0vf0, kernel ver >= 5.0 */
-	MLX5_PHYS_PORT_NAME_TYPE_UNKNOWN, /* Unrecognized. */
-};
+#include "mlx5_common.h"
 
-/** Switch information returned by mlx5_nl_switch_info(). */
-struct mlx5_nl_switch_info {
-	uint32_t master:1; /**< Master device. */
-	uint32_t representor:1; /**< Representor device. */
-	enum mlx5_nl_phys_port_name_type name_type; /** < Port name type. */
-	int32_t pf_num; /**< PF number (valid for pfxvfx format only). */
-	int32_t port_name; /**< Representor port name. */
-	uint64_t switch_id; /**< Switch identifier. */
-};
 
 /* VLAN netdev for VLAN workaround. */
 struct mlx5_nl_vlan_dev {
