+++ drivers/net/mlx5/mlx5_vlan.c
@@ -33,11 +33,11 @@
 
 #include <mlx5_glue.h>
 #include <mlx5_devx_cmds.h>
+#include <mlx5_nl.h>
 
 #include "mlx5.h"
 #include "mlx5_autoconf.h"
 #include "mlx5_rxtx.h"
-#include "mlx5_nl.h"
 #include "mlx5_utils.h"
 
 /**
